requirements.txt
streamlit
openai-whisper
yt-dlp
google-generativeai
fpdf
streamlit-markmap
groq
torch
numpy